+++ test/url_to_markdown/configuration_test.rb
@@ -39,9 +39,9 @@ class UrlToMarkdownConfigurationTest < Minitest::Test
     assert_raises(UrlToMarkdown::MissingCredentialsError) { config.cloudflare_account_id! }
   end
 
-  def test_default_actions_defaults_to_nil
+  def test_default_scripts_defaults_to_nil
     config = UrlToMarkdown::Configuration.new
 
-    assert_nil config.default_actions
+    assert_nil config.default_scripts
   end
 end